ch4: refactor vci init/finalize
Move multiple vci related init/finalize code into ch4_vci.c. Wrap
per-vci code into a function and only deal with vci 0 in ch4_init.c and
additional vcis in ch4_vci.c.
