+++ src/mpid/ch4/src/Makefile.mk
@@ -51,6 +51,7 @@ mpi_core_sources += src/mpid/ch4/src/ch4_globals.c        \
                     src/mpid/ch4/src/ch4_proc.c           \
                     src/mpid/ch4/src/ch4_stream_enqueue.c \
                     src/mpid/ch4/src/ch4_persist.c \
+                    src/mpid/ch4/src/ch4_vci.c \
 		    src/mpid/ch4/src/mpidig_init.c \
                     src/mpid/ch4/src/mpidig_recvq.c \
                     src/mpid/ch4/src/mpidig_pt2pt_callbacks.c \
