+++ src/mpid/ch4/src/ch4_impl.h
@@ -13,6 +13,10 @@
 #include "ch4_self.h"
 #include "ch4_vci.h"
 
+int MPIDI_vci_init(void);
+int MPIDI_vci_finalize(void);
+int MPIDI_init_per_vci(int vci);
+int MPIDI_destroy_per_vci(int vci);
 int MPIDIU_Intercomm_map_bcast_intra(MPIR_Comm * local_comm, int local_leader, int *remote_size,
                                      int *is_low_group, int pure_intracomm,
                                      int *remote_upid_size, char *remote_upids,
